Move `getMidToTrackId` function to a separate file
